Added roles, permissions, user_role, permission_role table and their respective relationships, defined a gate in authserviceprovider

+++ app/Http/Controllers/Api/PostController.php
@@ -17,6 +17,9 @@ class PostController extends Controller
      */
     public function index(Request $request)
     {
+        // gate usage, defined in AuthServiceProvider
+        $this->authorize('post_view');
+
         $orderCol = $request->input('order_column', 'id');
         $orderDir = $request->input('order_direction', 'desc');
 
@@ -60,6 +63,8 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
+        $this->authorize('post_create');
+
         $post = Post::create($request->validated());
 
         if ($request->hasFile('thumbnail')) {
@@ -78,6 +83,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('post_view');
         return new PostResource($post);
     }
 
@@ -90,6 +96,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        $this->authorize('post_update');
         $post->update($request->validated());
         return new PostResource($post);
     }
@@ -102,6 +109,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('post_delete');
         $post->delete();
 
         return response()->noContent();
